# Conflicts:
#	src/main/java/io/github/mzmine/gui/mainwindow/MainMenu.fxml

+++ src/main/java/io/github/mzmine/modules/dataprocessing/featdet_targeted/Gap.java
@@ -18,18 +18,18 @@
 
 package io.github.mzmine.modules.dataprocessing.featdet_targeted;
 
-import io.github.mzmine.datamodel.features.FeatureListRow;
-import io.github.mzmine.datamodel.features.ModularFeature;
-import java.util.List;
-import java.util.Vector;
 import com.google.common.collect.Range;
 import io.github.mzmine.datamodel.DataPoint;
 import io.github.mzmine.datamodel.FeatureStatus;
 import io.github.mzmine.datamodel.RawDataFile;
 import io.github.mzmine.datamodel.Scan;
+import io.github.mzmine.datamodel.features.FeatureListRow;
+import io.github.mzmine.datamodel.features.ModularFeature;
 import io.github.mzmine.datamodel.features.ModularFeatureList;
 import io.github.mzmine.datamodel.impl.SimpleDataPoint;
 import io.github.mzmine.util.scans.ScanUtils;
+import java.util.List;
+import java.util.Vector;
 
 class Gap {
 
@@ -175,7 +175,7 @@ class Gap {
         double intensityEnd = bestPeakDataPoints.get(i + 1).getIntensity();
 
         // calculate area of the interval
-        area += (rtDifference * (intensityStart + intensityEnd) / 2);
+        area += (float)(rtDifference * (intensityStart + intensityEnd) / 2);
 
       }
